// src/features/coin-tracker/domain/errors/coin-tracker-error-code.enum.ts
export enum CoinTrackerErrorCode {
  COIN_NOT_FOUND = 'COIN_NOT_FOUND',
  PRICE_ALERT_NOT_FOUND = 'PRICE_ALERT_NOT_FOUND',
  PRICE_ALERT_EXPIRED = 'PRICE_ALERT_EXPIRED',
  PRICE_ALERT_CANCELLED = 'PRICE_ALERT_CANCELLED',
  PRICE_ALERT_TRIGGERED = 'PRICE_ALERT_TRIGGERED',
  INVALID_EXPIRATION = 'INVALID_EXPIRATION',
  EMPTY_UPDATE = 'EMPTY_UPDATE',
  INVALID_CURSOR = 'INVALID_CURSOR',
  COINGECKO_API_ERROR = 'COINGECKO_API_ERROR'
}